+++ CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `RaainApiRainsProgressResponse.inQueue` field to track queued jobs count
 
+### Added
+
+- RainComputationCumulative windowInMinutes
+
 ## [3.1.10] - 2025-12-08
 
 ### Added